Split index.js into core components
