+++ components/copier_nonlooper.js
@@ -1,10 +1,11 @@
-var fbp = require('..');
+var InputPort = require('../core/InputPort')
+  , OutputPort = require('../core/OutputPort')
 
 // same as copier, but written as a non-looper
 
 module.exports = function copier_nonlooper() {
-  var inport = fbp.InputPort.openInputPort('IN');
-  var outport = fbp.OutputPort.openOutputPort('OUT');
+  var inport = InputPort.openInputPort('IN');
+  var outport = OutputPort.openOutputPort('OUT');
   // while (true) {
     var ip = inport.receive();
     // if (ip == null) {
