+++ components/reader.js
@@ -1,25 +1,28 @@
-var fbp = require('..');
-var Fiber = require('fibers');
-var fs = require('fs');
+var fbp = require('..')
+  , Fiber = require('fibers')
+  , fs = require('fs')
+  , InputPort = require('../core/InputPort')
+  , IP = require('../core/IP')
+	, OutputPort = require('../core/OutputPort');
 
  // Reader based on Bruno Jouhier's code
 
 module.exports = function reader() {
   var proc = fbp.getCurrentProc();
-  var inport = fbp.InputPort.openInputPort('FILE');
+  var inport = InputPort.openInputPort('FILE');
   var ip = inport.receive();
   var fname = ip.contents;
-  fbp.IP.drop(ip);
+  IP.drop(ip);
   fbp.setCallbackPending(true);
 
   var data = myReadFile(fname, "utf8", proc);
   console.log('read complete: ' + proc.name);
 
   fbp.setCallbackPending(false);
-  var outport = fbp.OutputPort.openOutputPort('OUT');
+  var outport = OutputPort.openOutputPort('OUT');
   var array = data.split('\n');
   for (var i = 0; i < array.length; i++) {
-    var ip = fbp.IP.create(array[i]);
+    var ip = IP.create(array[i]);
     outport.send(ip);
   }
 }
