+++ components/rrmerge.js
@@ -1,8 +1,9 @@
-var fbp = require('..');
+var InputPortArray = require('../core/InputPortArray')
+  , OutputPort = require('../core/OutputPort');
 
 module.exports = function rrmerge() {
-  var array = fbp.InputPortArray.openInputPortArray('IN');
-  var outport = fbp.OutputPort.openOutputPort('OUT');
+  var array = InputPortArray.openInputPortArray('IN');
+  var outport = OutputPort.openOutputPort('OUT');
   var ip = null;
   while (true) {
     for (var i = 0; i < array.length; i++) {
