+++ examples/fbptest8.js
@@ -1,4 +1,5 @@
-var fbp = require('..');
+var fbp = require('..')
+  , path = require('path');
 
 // --- define network ---
 var reader   = fbp.defProc('./components/reader');
@@ -6,7 +7,7 @@ var reverse  = fbp.defProc('./components/reverse');
 var reverse2 = fbp.defProc('./components/reverse', 'reverse2');
 var recvr    = fbp.defProc('./components/recvr');
 
-fbp.initialize(reader, 'FILE', './examples/data/text.txt');
+fbp.initialize(reader, 'FILE', path.resolve(__dirname, 'data/text.txt'));
 fbp.connect(reader, 'OUT', reverse, 'IN', 5);
 fbp.connect(reverse, 'OUT', reverse2, 'IN', 5);
 fbp.connect(reverse2, 'OUT', recvr, 'IN', 1);
